requirements.txt
PyQt6>=6.6.0
python-registry>=1.3.1